implements redirecting for specific error (statusCode) fixes https://github.com/dwyl/hapi-error/issues/5
