+++ lib/index.js
@@ -2,20 +2,25 @@
  * register defines our error_handler plugin
  */
 exports.register = function error_handler (server, options, next) {
-  // onPreResponse intercepts all errors
+  // onPreResponse intercepts ALL errors
   server.ext('onPreResponse', function (request, reply) {
     var req = request.response;
-    // console.log(request.response.message);
-    // console.log(req);
+
     if (req.isBoom) {
       var accept = request.raw.req.headers.accept;
       var statusCode = req.output.payload.statusCode;
-      // console.log(statusCode, typeof statusCode);
-      if(accept && accept.match(/json/)) { // support REST/JSON requests
+
+      // custom redirect https://github.com/dwyl/hapi-error/issues/5
+      if(options && options[statusCode] && options[statusCode].redirect) {
+        return reply.redirect(options[statusCode].redirect
+          + '?redirect=' + request.url.path).code(statusCode);
+      }
+      else if (accept && accept.match(/json/)) { // support REST/JSON requests
         return reply(req.output.payload).code(statusCode);
-      } else { // this is the default error message:
+      }
+      else { // this is the default error message:
         var msg = 'Sorry, something went wrong, please retrace your steps.';
-        if(req.message && req.isDeveloperError) {
+        if (req.message && req.isDeveloperError) {
           msg = req.message.replace('Uncaught error: ', '');
         }
         switch(statusCode) {
