+++ package.json
@@ -32,7 +32,7 @@
   "dependencies": {},
   "devDependencies": {
     "boom": "^3.1.3",
-    "decache": "^4.0.0",
+    "decache": "^4.1.0",
     "handlebars": "^4.0.5",
     "hapi": "^13.5.0",
     "hoek": "^4.0.0",
